move the contents of xgboost.parser into xgboost.trees
